Reorganizes Buffer\Tests tests.
